Refactoring(rename reducers)
