+++ project/src/store/user-data-reducer/selectors.ts
@@ -2,4 +2,4 @@ import {NameSpace} from '../root-reducer';
 import {State} from '../../types/state';
 import {AuthorizationStatus} from '../../const/authorization-status';
 
-export const getAuthorizationStatus = (state: State): AuthorizationStatus => state[NameSpace.dataUser].authorizationStatus;
+export const getAuthorizationStatus = (state: State): AuthorizationStatus => state[NameSpace.userData].authorizationStatus;
